backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
langchain-groq==0.2.2
langgraph==0.2.60
psycopg2-binary==2.9.10
langchain-core==0.3.25
